// src/api/generated/schemas/createCiRequest.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * cid-api
 * LDAP 로그인 + JWT 인가 API
 * OpenAPI spec version: v1
 */
import type { CiAppData } from './ciAppData';
import type { CiServerData } from './ciServerData';
import type { CiCloudData } from './ciCloudData';
import type { CiMiddlewareData } from './ciMiddlewareData';
import type { CiDatabaseData } from './ciDatabaseData';
import type { CiNetworkData } from './ciNetworkData';
import type { CiSecurityData } from './ciSecurityData';
import type { CiStorageData } from './ciStorageData';
import type { CiPcData } from './ciPcData';
import type { CiVdiData } from './ciVdiData';

export interface CreateCiRequest {
  /**
   * CI 명
   * @minLength 0
   * @maxLength 100
   */
  ciNm: string;
  /**
   * CI 타입 코드
   * @minLength 0
   * @maxLength 20
   */
  ciTpCd: string;
  /**
   * 업무 명
   * @minLength 0
   * @maxLength 30
   */
  ciBizwrkNm?: string;
  /**
   * CI 역할 명
   * @minLength 0
   * @maxLength 30
   */
  ciRoleNm?: string;
  /**
   * 환경 그룹 (PROD/STAGE/DEV)
   * @minLength 0
   * @maxLength 10
   */
  envrnGpCd?: string;
  /**
   * 등급 코드 (A/B/C)
   * @minLength 0
   * @maxLength 10
   */
  grdCd?: string;
  /** LOCATION ID */
  locId?: number;
  /** CI 설명 */
  ciDescp?: string;
  appData?: CiAppData;
  serverData?: CiServerData;
  cloudData?: CiCloudData;
  middlewareData?: CiMiddlewareData;
  databaseData?: CiDatabaseData;
  networkData?: CiNetworkData;
  securityData?: CiSecurityData;
  storageData?: CiStorageData;
  pcData?: CiPcData;
  vdiData?: CiVdiData;
}