// src/api/generated/schemas/ciServerData.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * cid-api
 * LDAP 로그인 + JWT 인가 API
 * OpenAPI spec version: v1
 */

export interface CiServerData {
  hostNm?: string;
  assetId?: string;
  ossId?: string;
  sysVidId?: string;
  deviceNm?: string;
  vendorId?: number;
  modelNm?: string;
  serialNo?: string;
  osTpNm?: string;
  osVer?: string;
  cpucoreCnt?: number;
  memoryCapa?: number;
  diskCapa?: number;
  virtMchnYn?: string;
  virtMchnPltfomNm?: string;
  rackId?: number;
  introDt?: string;
  maintEndDt?: string;
  monitYn?: string;
  osBackupYn?: string;
  alarmCallYn?: string;
  mngYn?: string;
  aciLvlGrd?: string;
  inetFacingYn?: string;
}